Rename chatbot and check for exit condition
